See merge request s07-blockchain-nft-sub2/S07P22A607!25

+++ backend/src/main/java/com/trippiece/backend/api/domain/dto/DistinctStickerDto.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 @Getter
 public class DistinctStickerDto implements Comparable<DistinctStickerDto> {
     private long id;
-    private int tokenId;
+    private long tokenId;
     private String tokenName;
     private String tokenURL;
     private int amount;
